+++ src/main/java/dev/lambdaurora/mcdev/api/manifest/ModBase.java
@@ -74,6 +74,7 @@ public class ModBase<SELF extends ModBase<SELF>> extends ModShell<SELF> {
 	public void copyTo(@NotNull ModBase<?> target) {
 		this.copyTo((ModShell<?>) target);
 		target.authors.addAll(this.authors);
+		target.contributors.addAll(this.contributors);
 		target.contact = this.contact != null ? this.contact.copy() : null;
 		target.license = this.license;
 	}